requirements.txt
altair==6.0.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
Bottleneck @ file:///opt/miniconda3/conda-bld/bottleneck_1761937849374/work
cachetools==6.2.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
delta-spark==3.1.0
exceptiongroup==1.3.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kafka-python==2.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.16.0
nest-asyncio==1.6.0
numexpr @ file:///opt/miniconda3/conda-bld/numexpr_1762165517928/work
numpy @ file:///opt/miniconda3/conda-bld/numpy_and_numpy_base_1761579345363/work/dist/numpy-2.2.5-cp310-cp310-macosx_12_0_arm64.whl#sha256=cee7bef98ac83b78786789b80f44469231df1e3cf12ce327f909aa4ea0fcee88
packaging==25.0
pandas @ file:///opt/miniconda3/conda-bld/pandas_1762332182781/work/dist/pandas-2.3.3-cp310-cp310-macosx_12_0_arm64.whl#sha256=15e77fd5caae22506c552ef9e15f2690708d3395500d8e3ac8ce5c5cf227c7a1
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pyspark==3.5.7
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
python-dotenv==1.2.1
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3bjczh6it0/croot/pytz_1752135865541/work
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
smmap==5.0.2
stack-data==0.6.3
streamlit==1.54.0
tenacity==9.1.3
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.6.2
wcwidth==0.2.14
zipp==3.23.0
